+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+7.9.3
+=====
+
+*   (improvement) Add dynamic default for `$status` in `AjaxResponseBuilder`.
+*   (improvement) Make `LabeledExceptionTrait::setFrontendMessage()` fluent.
+*   (improvement) Add `LabeledExceptionTrait::createWithLabel()`.
+
+
 7.9.2
 =====